requirements.txt
asgiref==3.7.2
blinker==1.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.5
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
filelock==3.6.0
gyp==0.1
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
Pillow==10.0.0
platformdirs==2.5.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
pytz==2023.3.post1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.13.0+ds
wadllib==1.3.6
zipp==1.0.0
